NetService removal refactoring
Moved NetService GetSegmentLaneIdsAndLaneIndexes into ExtSegmentManager.
Moved GetSegmentLaneIdsEnumerable into TrafficManager.Manager.Impl namespace.
Moved GetSegmentLaneIdsEnumerator into TrafficManager.Manager.Impl namespace.
Moved LaneIdAndIndex TrafficManager.Manager.Impl namespace.
