+++ TLM/TLM/TLM.csproj
@@ -147,6 +147,9 @@
     <Compile Include="Manager\Impl\GeometryNotifier.cs" />
     <Compile Include="Manager\Impl\GetNodeSegmentIdsEnumerable.cs" />
     <Compile Include="Manager\Impl\GetNodeSegmentIdsEnumerator.cs" />
+    <Compile Include="Manager\Impl\GetSegmentLaneIdsEnumerable.cs" />
+    <Compile Include="Manager\Impl\GetSegmentLaneIdsEnumerator.cs" />
+    <Compile Include="Manager\Impl\LaneIdAndIndex.cs" />
     <Compile Include="Notifier.cs" />
     <Compile Include="Patch\CustomPathFindPatchAttribute.cs" />
     <Compile Include="Patch\PatchCommons.cs" />
