+++ TLM/TLM/Traffic/Impl/SegmentEnd.cs
@@ -318,7 +318,8 @@ namespace TrafficManager.Traffic.Impl {
             numVehiclesMovingToSegmentId = new Dictionary<ushort, uint>[numLanes];
             numVehiclesGoingToSegmentId = new Dictionary<ushort, uint>[numLanes];
 
-            foreach (LaneIdAndIndex laneIdAndIndex in NetService.Instance.GetSegmentLaneIdsAndLaneIndexes(SegmentId)) {
+            ExtSegmentManager extSegmentManager = ExtSegmentManager.Instance;
+            foreach (LaneIdAndIndex laneIdAndIndex in extSegmentManager.GetSegmentLaneIdsAndLaneIndexes(SegmentId)) {
                 var numVehicleMoving = new Dictionary<ushort, uint>();
                 var numVehicleGoing = new Dictionary<ushort, uint>();
 
@@ -327,7 +328,6 @@ namespace TrafficManager.Traffic.Impl {
             }
 
             IExtSegmentEndManager segEndMan = Constants.ManagerFactory.ExtSegmentEndManager;
-            ExtSegmentManager extSegmentManager = ExtSegmentManager.Instance;
             for (int i = 0; i < 8; ++i) {
                 ushort segId = node.GetSegment(i);
                 if (segId == 0) {
