+++ TLM/TLM/Util/Record/LaneConnectionRecord.cs
@@ -93,7 +93,7 @@ namespace TrafficManager.Util.Record {
                     continue;
                 }
 
-                foreach (LaneIdAndIndex laneIdAndIndex in NetService.Instance.GetSegmentLaneIdsAndLaneIndexes(segmentId)) {
+                foreach (LaneIdAndIndex laneIdAndIndex in extSegmentManager.GetSegmentLaneIdsAndLaneIndexes(segmentId)) {
                     NetInfo.Lane laneInfo = netInfo.m_lanes[laneIdAndIndex.laneIndex];
                     bool match = (laneInfo.m_laneType & LaneConnectionManager.LANE_TYPES) != 0 &&
                                  (laneInfo.m_vehicleType & LaneConnectionManager.VEHICLE_TYPES) != 0;
