+++ TLM/TMPE.CitiesGameBridge/Service/NetService.cs
@@ -12,19 +12,6 @@ namespace CitiesGameBridge.Service {
 
         // OTHER STUFF --------------------------------------------------------------------------------
 
-        public GetSegmentLaneIdsEnumerable GetSegmentLaneIdsAndLaneIndexes(ushort segmentId) {
-            NetManager netManager = Singleton<NetManager>.instance;
-            ref NetSegment netSegment = ref netManager.m_segments.m_buffer[segmentId];
-            uint initialLaneId = netSegment.m_lanes;
-            NetInfo netInfo = netSegment.Info;
-            NetLane[] laneBuffer = netManager.m_lanes.m_buffer;
-            if (netInfo == null) {
-                return new GetSegmentLaneIdsEnumerable(0, 0, laneBuffer);
-            }
-
-            return new GetSegmentLaneIdsEnumerable(initialLaneId, netInfo.m_lanes.Length, laneBuffer);
-        }
-
         public IList<LanePos> GetSortedLanes(ushort segmentId,
                                              ref NetSegment segment,
                                              bool? startNode,
