+++ TLM/TMPE.GenericGameBridge/TMPE.GenericGameBridge.csproj
@@ -75,11 +75,8 @@
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
     <Compile Include="Factory\IServiceFactory.cs" />
-    <Compile Include="Service\GetSegmentLaneIdsEnumerable.cs" />
-    <Compile Include="Service\GetSegmentLaneIdsEnumerator.cs" />
     <Compile Include="Service\INetService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Service\LaneIdAndIndex.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
